[kotlin] k2 Completion: fix missing suggestions for second and further named argument in case of private function
Another possible solution would be to use `analyzeCopy` with `PREFER_SELF` in `KotlinFirCompletionProvider`, but this way duplicate symbols (one from the original file, and another one from the copy) start appearing in a lot of cases. They come from the package scope and are not easily filtered out. Also, some of them might not be covered by tests. So it would be more preferable to use `analyzeCopy` with `IGNORE_ORIGIN`, and not `PREFER_SELF`, in `KotlinFirCompletionProvider`, which will be possible once KT-68929 is implemented. For now, use a workaround by adding `analyzeCopy` with `PREFER_SELF`
to `FirNamedArgumentCompletionContributor` specifically.

 ^KTIJ-29748 Fixed

GitOrigin-RevId: 1a33b3ebfcb198b7f4fc0160cf0176474da99de2