[gradle][kotlin] IDEA-354292 fix performance warnings in logs because of creating a gutter button for a non-leaf element
There were a lot of warnings in logs because of creating a run button in the gutter for KtStringTemplate element:
```
WARN - #c.i.c.d.LineMarkerInfo - Performance warning: LineMarker is supposed to be registered for leaf elements only, but got: STRING_TEMPLATE (class org.jetbrains.kotlin.psi.KtStringTemplateExpression) instead. First child: PsiElement(OPEN_QUOTE) (class com.intellij.psi.impl.source.tree.LeafPsiElement)
Please see LineMarkerProvider#getLineMarkerInfo(PsiElement) javadoc for detailed explanations.
```
It was happening because KtStringTemplateExpression is not a LeafPsiElement. Firstly, I chose this type because I was looking for the most precise element that could contain a task name. It could contain a lot of LeafPsiElement children. For each of them `KotlinGradleTaskRunConfigurationProducer` could produce the same run configuration for a given task name. But `KotlinGradleTaskRunLineMarkerProvider` should create Info only for the one element at the line to avoid the duplication of run configuration options in the gutter.

To fix these warnings, it's needed to use a child of string template expression which is not duplicating. I chose an open quote element for that.

GitOrigin-RevId: a822f21b01c0e7da6868ce46d1247b9358cf91a8