Add K2 compose checker tests
So far, we ran smoke tests for compose checker on Android Studio, but
now the compose compiler plugin is a part the upstream KT compiler, and
we enabled it for the upstream IntelliJ as well. This commit adds
similar smoke tests to IntelliJ. Since we will work on the compose
support for KMP, I believe this new smock test helps us to detect some
issues of enabling the compose compiler plugin checker.

^KTIJ-30601

GitOrigin-RevId: 6c461ac9435d180b5cdda2529a400a0da4b8aa6c