KTIJ-26423 [kotlin] Implement basic version of `OptimizedImportsBuilder`
`OptimizedImportsBuilder` is mostly identical to the
`OptimizedImportsBuilder` from the K1 implementation
of Import Optimizer.

It does not have conflicts resolution yet.

This makes the K2 implementation of Import Optimizer
to adhere to the Import Layout settings.

It also fixes the problem with default Kotlin
imports not being removed on performing the
import optimization (KTIJ-26751).

^KTIJ-26423 Fixed
^KTIJ-26751 Fixed

GitOrigin-RevId: 397349af768dc12a2ca1796f316c9240bf5fb2ec