KTIJ-30883 [kotlin] Use different checks for static and regular imported members in Import Optimizer
Regular non-static declarations are always dispatched with implicit
receivers; thus, it is enough to find a compatible type in the
implicit receivers list to check if the declaration
is indeed imported or just dispatched.

On the other hand, static declarations do not depend on implicit receivers at all;
instead, they come from scopes which can be introduced only by extending a
Java class with static declarations.
So we have to traverse the scopes to do the check.

^KTIJ-30883 Fixed

GitOrigin-RevId: e2efe818ed45d34e8d3f3778dcb280d575378df9