[kotlin debugger] IDEA-361702 Fix smart step into with empty constructor
Drop the code that previously excluded empty constructor from the list of smart step targets to fix the following scenario. Consider the case `consume(A())`, where `A()` is an empty constructor. If we exclude it from the targets, there will be only one target to suggest, so the smart step into machinery will fall back to a regular step into. But the constructor executes first, so the debugger suspends in it. As a result, we targeted to skip the constructor, but promoted it even more.

With this fix, the constructor is suggested, but the user can easily select the target they want with more control.

GitOrigin-RevId: dcf72978259fd51c3ab83774eb4c5da6e4d4a20b