[gradle][kotlin] IDEA-361677 add navigation tests "TOML <-> build.gradle.kts" for an included build without settings
The previous changes for IDEA-361677 reworked the logic of collecting version catalogs considering the case when an included build does not have a settings file. This commit adds tests to check this case for build scripts in Kotlin.

I also tried adding a test for navigation from `build.gradle.kts` of a subproject to a TOML of its parent multi-project build (by the subproject I mean the module that is specified in the root settings file by `include(":subproject")`). Unfortunately, this test was failing due to unsuccessful resolve of UAST element. But the test project was valid: when I opened it in IDEA, the desired navigation worked.

GitOrigin-RevId: 670946f7d3faa5bbe3aa7594a304c6b1908d1f27