KTIJ-32202 [kotlin] Do not lock out conflicting symbols for the type positions
Locking in `java.lang.Deprecated` also led to locking out `kotlin.Deprecated`
alternative symbol.
This, in turn, led to adding explicit imports for everything from `kotlin.*`
package.

This seems like a wrong logic, but this will be dealt with later,
probably as a part of KTIJ-32231.

The current solution is to not lock out the `kotlin.Deprecated`
symbol for `KtUserType` positions (it includes annotation calls).
This logic is already present in K1 Mode's `OptimizedImportsBuilder`.

Also, add a muted testcase for KTIJ-32231 (doesn't work in both K1 and K2)

Also, add muting logic to the K1's `AbstractOptimizeImportsTest`

^KTIJ-32202 Fixed

GitOrigin-RevId: 89ce6287a44bd05b55cc6bbb3b01157bee30b41c