[kotlin] Remove custom seq in favor of stdlib sequence
I found a bug in TransformingIterator: hasNext isn't side-effect-free,
and because of that calling hasNext() twice skips two elements.

But then, I found that in comment in review IJ-MR-107533,
where the custom seq was implemented, it was said that
this seq is a temporary solution.

So, let's remove it then.

^KTIJ-32260

GitOrigin-RevId: 0679789333698ec3a994ab05bd9f62fb75c2bc49