[kotlin] KotlinDiagnosticHighlightVisitor: optimize diagnostic reporter
The logic had a big algorithmic complexity, so more diagnostics lower performance.
`PsiElement` itself can be used as an anchor for
diagnostics to avoid search for some specific diagnostic
range.
This optimization is based on the fact that `visit` method
will iterate through all elements, so effectively it will
check all elements from `diagnosticsMap`.
`HighlightRangeExtension` implementation is required
to not skip parent diagnostics

^KTIJ-32379 Fixed

GitOrigin-RevId: 32eb9d091de184e66f1ec96da9be4d59df1a616c