Implement `FilterIsInstanceResultIsAlwaysEmptyInspection`
The inspection detects calls to `filterIsInstance`/`filterIsInstanceTo`, whose target type is not a subtype or supertype of the element type the filtering is called on.

KTIJ-28490 `Add inspections for useless filtering` contains a non-trivial counterexample,
which shows that just checking that the target type is a subtype of the source type is not enough.
That's the reason why the code includes a short proof of the inspection's correctness.

^KTIJ-28490 fixed

GitOrigin-RevId: 0e70841a5030330cf58effc7077c1a06c575ca54