[intellij project configuration] add missing '-Xjvm-default=all' parameter in Kotlin compiler arguments
All modules in intellij project use should use this parameter to ensure that newly added interface methods with default implementation will have it in bytecode to avoid breaking compatibility.

We have a test which verifies that, but it was disabled some time ago by a mistake. This change is also needed to re-enable it.

GitOrigin-RevId: 8fec5d419bf8da2b1fa7058d61ef04a3e8ebc6fa