[intellij project configuration] remove Kotlin facets which don't specify any custom options for a module
If there is no Kotlin facet in a module, it automatically inherits the project-wide options from .idea/kotlinc.xml. So if a configuration of a Kotlin facet in a module doesn't specify any custom options, the facet is not necessary, and it can be removed to simplify the project configuration and make it easier to modify project-wide options later.

Also, a test which reports such unnecessary Kotlin facets is added.

GitOrigin-RevId: 7cf2a714064499e1cdd768978ff7afd839567c87