[Kotlin] K1: remove outdated experimental feature quick fix
The only feature supported by the quick fix is 'inline classes'.
It is not experimental in Kotlin versions currently supported by
the Intellij Kotlin plugin.

The quick fix itself doesn't work as expected as it has
side effects in preview. But outside the test environment, it's not possible
to encounter the issue.

Relates to KTIJ-30918

GitOrigin-RevId: 23f2d4939b3519c2ed74e849a2ed160628ce4e8c