[kotlin] cleanup: get rid of fake packages in content modules of Kotlin Frontend plugin (KTIJ-32906)
Fake packages were used to ensure that the main plugin classloader is used to load the classes from content modules. Now the same can be achieved explicitly with loading="embedded" attribute. This is also needed to ensure that the classpath for content modules can be properly specified (this is needed for IJPL-165692).

GitOrigin-RevId: 94965f7cde9083e4b4f8eb460f0da94f004b7cc9