KTIJ-32842 [kotlin] Avoid lazy registration of auto-imports for delegation errors
It is not clear which
reference should be used in case of delegation.

The `DelegateSpecialFunction*` diagnostics point
to the delegate expression, and that expression can
be anything at all.
For example, a string literal does not have
a reference on it, so `KotlinDiagnosticHighlightVisitor`
would not be able to call `registerUnresolvedReferenceLazyQuickFixes`
for it.

From the performance perspective, this should be OK,
since the delegation happens not that often in the code.

GitOrigin-RevId: a1ac6297179be16f25c6ce3f67d5f77da81d7939