[kotlin.base.scripting] rename the Amper dependency resolution library
Amper libraries are updated in the Amper IDE plugin very frequently to keep the plugin as up-to-date as possible. But it is not necessary for the kotlin.base.scripting usages, there the lib can be updated in another pace. It is also not convenient for Amper developers to bump versions in the kotlin.base.scripting as well.

Since it is not allowed to have different versions of a single library in the IntelliJ project, the easiest way around this is to have different names for these libraries.

GitOrigin-RevId: 66d1b82768761c6cad9c9c309266734247f34747