[kotlin] tests: create KotlinSdk before the project is created
to ensure that the corresponding sdk entry is found in the project snapshot. Simplification of the fix 688c170cbb3e3a73265d78ecc0c535c29f62ed36 which is now works for all tests based on KotlinMultiPlatformProjectDescriptor.

fixes KotlinDefinitionsSearcherMultiplatformTestGenerated.KotlinCallable.testProperty if it ever tries to call analyze and build dependencies for KaSession

GitOrigin-RevId: 83c86d900c8e7a5e732878ee7eb3d49f14324ce4