[kotlin] Fix failing tests for Kotlin/JS import
The tests that were failing:
- org.jetbrains.kotlin.idea.codeInsight.gradle.GradleFacetImportTest8.testJsImportWithCustomSourceSets
- org.jetbrains.kotlin.idea.codeInsight.gradle.GradleFacetImportTest8.testJsImport
- org.jetbrains.kotlin.idea.codeInsight.gradle.GradleFacetImportTest8.testStableModuleNameWhileUsingGradleJS

The issue was that IDEA uses the `K2JSCompilerArguments` class
from the Kotlin compiler to parse CLI options passed by KGP
to build its project model.
The `-output` option was removed from that class, so normally
we'd just ignore it as unknown.
However, this option expects a value as the next argument.
So, after it was removed from `K2JSCompilerArguments`,
its value would be parsed as a free positional argument,
which was incorrect.

The `-output` CLI option has been restored in the Kotlin compiler,
and here we bring back the logic that ignores it.

GitOrigin-RevId: b2db8b732d5817c9e2d7ea2aeb096a3038207821