[debugger] Fix symmetricity of CoroutineStackFrame
It could be equal to any other JavaStackFrame or its descendant that has the same frame proxy. However, JavaStackFrame.equals ensures it can only be equal to another frame of the very same type. By missing delegation to this method, CoroutineStackFrame breaks the symmetricity of equality between itself and other types of JavaStackFrame.

GitOrigin-RevId: ffc5cb6f6c689c37efd99b5c76fde621c7589114