[kotlin] Add a rough implementation of `SuspiciousImplicitCoroutineScopeReceiverAccessInspection`
This inspection allows detecting the usages or `CoroutineScope`s
in a potentially different suspension context.

Note 1: This implementation is DISABLED by default and pushed to gather
the feedback regarding the current implementation, messages,
description, options and so on.

Note 2: Currently the implementation does not provide any quick fixes.
This is done explicitly, because it is not yet clear what would be
the best quickfix, and if adding an explicit receiver would be enough.

Related tickets: KTIJ-16542, KTIJ-12307

GitOrigin-RevId: ab69f291e8ecd4121f176f2c6865123984ede47d