KT-77210 [kotlin] Account for the changed behaviour of `ContextCollector` in `CodeFragmentContextTuner`
This is required for the debugger to be able to correctly step
into properties declared in primary constructors.

For such positions, `ContextCollector` would now collect
the scopes which are relevant for the constructor.

However, the debugger is interested in the members of the
class, and not in the constructor parameters.

So using the whole class as the context element
makes `ContextCollector` provide correct scopes
with the correct implicit receiver of the outer class,
which fixes the "Step Into" behaviour.

The failing test was:
InlineScopesAndK2IdeK2CodeEvaluateExpressionTestGenerated.SingleBreakpoint.Uncategorized.testOnGetter

GitOrigin-RevId: 4cb4bb1211b05724061e25d123af35ec7353c150