[kt, tests] Refactoring: remove versions from KotlinMppTestsContext
Expose 'testProperties' instead of fields like 'gradleVersion' in KotlinMppTestsContext.

Versions have a lot of complicated handling in KotlinMppTests (substitutions,
resolution from CI). This commit incapsulates all of that in KotlinTestProperties.

Why it's good? Because it makes `KotlinMppTestsContext` connected with KMP only
as much as KotlinTestProperties do (they are very KMP-aware at the moment, but
the next commit will change that)

GitOrigin-RevId: 3a6cef863d36e80f488bcf0889998144cad7a4e4