[platform ml] convert intellij.platform.ml.impl to a content module (IJPL-190495)
This is needed to have an explicit dependency on the backend in it (IJPL-188916).

The new module is registered in common-ide-modules.xml to include it in all IntelliJ-based IDEs as before. It's also explicitly included in Fleet Backend to keep the old behavior. Existing plugins which depend on intellij.platform.ml.impl are updated to have an explicit dependency. 'jetbrains.mlapi*' and 'ngram' libraries which are used by the module are also moved to its content module, and other plugins which use these libraries now do it via dependency on the module.

GitOrigin-RevId: be9dc19406b87f8978f7a922e950f9009e1aabda