[gradle][debugger] IDEA-343854 implement `SoftlyKillableProcessHandler` interface in `ExternalSystemProcessHandler`
Prevents forceful termination of a Gradle Daemon in `com.intellij.debugger.impl.DebuggerManagerImpl.attachVirtualMachine`, when debugging Gradle scripts is enabled. It is not the responsibility of `DebuggerManagerImpl` to stop the Gradle Daemon. `ExternalSystemProcessHandler` is much more suitable for handling Gradle application stopping. It sends the cancellation token to the Gradle Daemon. And then, the Daemon gracefully stops the processes created by itself. If the Daemon is killed by `DebuggerManagerImpl`, its child non-JVM processes are still alive and are detached to the root process (00001) (see the webpack example in comments for IDEA-343854).

Also, it seems incorrect to terminate the Gradle Daemon in `DebuggerManagerImpl` because it did happen only when debugging Gradle scripts was enabled in the Run Configuration. If it was disabled while debugging, or if a Run Configuration was executed without a debug at all - the Daemon was alive even after stopping the configuration. This commit brings more consistency because the Daemon will not be stopped in any of the above-mentioned cases.

FYI, there is a race condition found. Both `DebuggerManagerImpl` and `ExternalSystemProcessHandler` are called in `com.intellij.execution.process.ProcessHandler.destroyProcess`:
* DebuggerManagerImpl - in `fireProcessWillTerminate(true)`
* ExternalSystemProcessHandler in `destroyProcessImpl()`
They execute the process-stopping logic asynchronously relatively to each other. Usually, `DebuggerManagerImpl` tries to stop a process earlier.

GitOrigin-RevId: e5eb52eb6dbcc4e19d2d0632421114490b628e37